+++ bin/commands/add.js
@@ -1,21 +1,27 @@
 "use strict";
+var __spreadArrays = (this && this.__spreadArrays) || function () {
+    for (var s = 0, i = 0, il = arguments.length; i < il; i++) s += arguments[i].length;
+    for (var r = Array(s), k = 0, i = 0; i < il; i++)
+        for (var a = arguments[i], j = 0, jl = a.length; j < jl; j++, k++)
+            r[k] = a[j];
+    return r;
+};
 Object.defineProperty(exports, "__esModule", { value: true });
 exports.add = void 0;
-const schemes_1 = require("../schemes");
-const utils_1 = require("../utils");
-const config_1 = require("../utils/config");
-const schemes_2 = require("../utils/schemes");
-const add = (commanderConfig) => {
-    const { folder, schemes: schemesDir, extension, name, methods, test, } = config_1.setupCommand(commanderConfig);
-    const [userFiles, userSchemes] = schemesDir ? schemes_2.loadSchemes(schemesDir) : [];
-    const files = userFiles
-        ? utils_1.joinFiles(folder, ...utils_1.renameFiles(name, ...userFiles).map(file => file + '.' + extension))
-        : utils_1.joinFiles(folder, `index.${extension}`, `${name}.handlers.${extension}`, `${name}.test.${extension}`);
-    const schemes = userSchemes || [
+var schemes_1 = require("../schemes");
+var utils_1 = require("../utils");
+var config_1 = require("../utils/config");
+var schemes_2 = require("../utils/schemes");
+var add = function (commanderConfig) {
+    var _a = config_1.setupCommand(commanderConfig), folder = _a.folder, schemesDir = _a.schemes, extension = _a.extension, name = _a.name, methods = _a.methods, test = _a.test;
+    var _b = schemesDir ? schemes_2.loadSchemes(schemesDir) : [], userFiles = _b[0], userSchemes = _b[1];
+    var files = userFiles
+        ? utils_1.joinFiles.apply(void 0, __spreadArrays([folder], utils_1.renameFiles.apply(void 0, __spreadArrays([name], userFiles)).map(function (file) { return file + '.' + extension; }))) : utils_1.joinFiles(folder, "index." + extension, name + ".handlers." + extension, name + ".test." + extension);
+    var schemes = userSchemes || [
         schemes_1.routerSchema(name, methods),
         schemes_1.controllerSchema(name, methods),
         schemes_1.testSchema(name, methods),
     ];
-    utils_1.generate({ files, schemes }, { position: files.length - 1, condition: test });
+    utils_1.generate({ files: files, schemes: schemes }, { position: files.length - 1, condition: test });
 };
 exports.add = add;