+++ bin/commands/remove.js
@@ -1,21 +1,22 @@
 "use strict";
 Object.defineProperty(exports, "__esModule", { value: true });
 exports.remove = void 0;
-const fs_1 = require("fs");
-const path_1 = require("path");
-const constants_1 = require("../constants");
-const remove = ({ name, path }) => {
-    const { rootDir } = constants_1.loadConfig();
+var fs_1 = require("fs");
+var path_1 = require("path");
+var constants_1 = require("../constants");
+var remove = function (_a) {
+    var name = _a.name, path = _a.path;
+    var rootDir = constants_1.loadConfig().rootDir;
     if (!path)
         path = rootDir;
-    const folder = path_1.join(path, name);
+    var folder = path_1.join(path, name);
     if (fs_1.existsSync(folder)) {
-        fs_1.rm(folder, { recursive: true, force: true }, () => {
+        fs_1.rm(folder, { recursive: true, force: true }, function () {
             console.log('removed', name);
         });
     }
     else {
-        console.log(`didn't find folder '${name}' under ${path}`);
+        console.log("didn't find folder '" + name + "' under " + path);
     }
 };
 exports.remove = remove;