+++ bin/constants.js
@@ -1,9 +1,9 @@
 "use strict";
 Object.defineProperty(exports, "__esModule", { value: true });
 exports.loadConfig = exports.errorCase = exports.fileName = void 0;
-const path_1 = require("path");
+var path_1 = require("path");
 exports.fileName = 'erg.config.js';
-const errorCase = (err, file) => {
+var errorCase = function (err, file) {
     if (err.code === 'MODULE_NOT_FOUND') {
         console.error(file, 'is missing!');
     }
@@ -12,7 +12,7 @@ const errorCase = (err, file) => {
     }
 };
 exports.errorCase = errorCase;
-const loadConfig = () => {
+var loadConfig = function () {
     try {
         return require(path_1.join(process.cwd(), exports.fileName));
     }