+++ bin/index.js
@@ -1,11 +1,11 @@
 #!/usr/bin/env node
 "use strict";
 Object.defineProperty(exports, "__esModule", { value: true });
-const commander_1 = require("commander");
-const commands_1 = require("./commands");
-const constants_1 = require("./constants");
-const { version, description } = require('../package.json');
-const program = new commander_1.Command();
+var commander_1 = require("commander");
+var commands_1 = require("./commands");
+var constants_1 = require("./constants");
+var _a = require('../package.json'), version = _a.version, description = _a.description;
+var program = new commander_1.Command();
 program.version(version, '-v, --version');
 program.description(description);
 program
@@ -18,10 +18,11 @@ program
     .option('-s, --schemes <path>', 'path of the schemes root folder')
     .option('-m, --methods <methods...>', 'accepted methods')
     .option('--typescript', 'generates the files with .ts extension', false)
-    .option('--no-test', 'prevents generation of test file', false)
-    .action((name, { path, schemes, methods, typescript, test }, _) => {
+    .option('--no-test', 'prevents generation of test file')
+    .action(function (name, _a, _) {
+    var path = _a.path, schemes = _a.schemes, methods = _a.methods, typescript = _a.typescript, test = _a.test;
     try {
-        commands_1.add({ name, path, schemes, methods, typescript, test });
+        commands_1.add({ name: name, path: path, schemes: schemes, methods: methods, typescript: typescript, test: test });
     }
     catch (err) {
         constants_1.errorCase(err, constants_1.fileName);
@@ -34,9 +35,10 @@ program
     name: 'name of the route to be removed',
 })
     .option('-p, --path <path>', 'path of the routes root folder')
-    .action((name, { path }, _) => {
+    .action(function (name, _a, _) {
+    var path = _a.path;
     try {
-        commands_1.remove({ name, path });
+        commands_1.remove({ name: name, path: path });
     }
     catch (err) {
         constants_1.errorCase(err, constants_1.fileName);