+++ bin/schemes/router.schema.js
@@ -1,13 +1,8 @@
 "use strict";
 Object.defineProperty(exports, "__esModule", { value: true });
 exports.routerSchema = void 0;
-const schemes_1 = require("../utils/schemes");
-const routerSchema = (name, methods) => `import { Router } from 'express';
-${schemes_1.imports(name, methods)}
-  
-const router = Router();
-  
-router.route('/')${schemes_1.controllersRef(methods)};
-  
-export default router;`;
+var schemes_1 = require("../utils/schemes");
+var routerSchema = function (name, methods) {
+    return "import { Router } from 'express';\n" + schemes_1.imports(name, methods) + "\n  \nconst router = Router();\n  \nrouter.route('/')" + schemes_1.controllersRef(methods) + ";\n  \nexport default router;";
+};
 exports.routerSchema = routerSchema;