+++ bin/utils/schemes/controllers.js
@@ -1,9 +1,15 @@
 "use strict";
 Object.defineProperty(exports, "__esModule", { value: true });
 exports.controllersRef = exports.controllers = void 0;
-const controllers = (methods) => methods.map(method => addController(method)).join('\n');
+var controllers = function (methods) {
+    return methods.map(function (method) { return addController(method); }).join('\n');
+};
 exports.controllers = controllers;
-const controllersRef = (methods) => methods.map(method => addControllerRef(method)).join('');
+var controllersRef = function (methods) {
+    return methods.map(function (method) { return addControllerRef(method); }).join('');
+};
 exports.controllersRef = controllersRef;
-const addController = (method) => `export const ${method}Controller = async (req, res, next) => {};`;
-const addControllerRef = (method) => `.${method}(${method}Controller)`;
+var addController = function (method) {
+    return "export const " + method + "Controller = async (req, res, next) => {};";
+};
+var addControllerRef = function (method) { return "." + method + "(" + method + "Controller)"; };