+++ bin/utils/schemes/imports.js
@@ -1,6 +1,10 @@
 "use strict";
 Object.defineProperty(exports, "__esModule", { value: true });
 exports.imports = void 0;
-const addImports = (methods) => methods.map(method => `${method}Controller`).join(', ');
-const imports = (name, methods) => `import { ${addImports(methods)} } from './${name}.handlers';`;
+var addImports = function (methods) {
+    return methods.map(function (method) { return method + "Controller"; }).join(', ');
+};
+var imports = function (name, methods) {
+    return "import { " + addImports(methods) + " } from './" + name + ".handlers';";
+};
 exports.imports = imports;