+++ bin/utils/schemes/test.js
@@ -1,12 +1,10 @@
 "use strict";
 Object.defineProperty(exports, "__esModule", { value: true });
 exports.tests = void 0;
-const tests = (name, methods) => methods.map(method => addTest(name, method)).join('\n  ');
+var tests = function (name, methods) {
+    return methods.map(function (method) { return addTest(name, method); }).join('\n  ');
+};
 exports.tests = tests;
-const addTest = (name, method) => `it('${method}', async done => {
-    const response = await supertest(server).get('/${name}');
-    
-    // expectations
-
-    done();
-  });`;
+var addTest = function (name, method) {
+    return "it('" + method + "', async done => {\n    const response = await supertest(server).get('/" + name + "');\n    \n    // expectations\n\n    done();\n  });";
+};