+++ CHANGELOG.md
@@ -1,5 +1,6 @@
-# Changelog v1.0.0-release
+# Changelog v1.0.0
 
 ## Fixed
 
-- bug where cli couldn't operate without custom schemes
+- bug where cli couldn't interpret `language` field
+- bug where cli couldn't interpret `test` field